[REFACTOR]: item에서 select 하는 부분 findItem으로 변경
