+++ pickme/src/main/java/project/pickme/item/repository/FindItemMapper.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import project.pickme.item.domain.Item;
 import project.pickme.item.dto.FindItemDto;
 
 @Mapper
@@ -13,4 +14,6 @@ public interface FindItemMapper {
 	Optional<FindItemDto.GetOne> findById(@Param("id") Long id, @Param("userId") String currentUserId);
 
 	List<FindItemDto.GetAll> findAll();
+
+	Optional<Item> findItemById(Long itemId);
 }
