+++ pickme/src/main/java/project/pickme/item/repository/ItemMapper.java
@@ -1,17 +1,11 @@
 package project.pickme.item.repository;
 
 import java.util.List;
-import java.util.Optional;
-
 import org.apache.ibatis.annotations.Mapper;
-
-import project.pickme.item.domain.Item;
 import project.pickme.item.dto.ItemDto;
 
 @Mapper
 public interface ItemMapper {
-	Optional<Item> findItemById(Long itemId);
-
 	void insertItem(ItemDto itemDto);
 
 	void deleteAll();
