+++ pickme/src/main/java/project/pickme/item/service/FindItemService.java
@@ -1,4 +1,6 @@
-package project.pickme.item;
+package project.pickme.item.service;
+
+import static project.pickme.item.exception.ItemErrorCode.*;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -6,8 +8,12 @@ import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
+import project.pickme.bid.dto.response.OneBidItemDto;
+import project.pickme.common.exception.BusinessException;
+import project.pickme.item.domain.Item;
 import project.pickme.item.dto.FindItemDto;
 import project.pickme.item.repository.FindItemMapper;
+import project.pickme.user.domain.User;
 
 @Service
 @RequiredArgsConstructor
@@ -23,5 +29,12 @@ public class FindItemService {
 
 	public List<FindItemDto.GetAll> findAll() {
 		return itemMapper.findAll();
+	}
+
+	public OneBidItemDto showOneBidItem(User user, Long itemId) {
+		Item item = itemMapper.findItemById(itemId).orElseThrow(() -> new BusinessException(NOT_FOUND_ITEM));
+
+		//TODO: 이미지 조회 로직
+		return OneBidItemDto.createOf(item, user,"test.png");
 	}
 }
