+++ pickme/src/main/java/project/pickme/s3/service/S3Service.java
@@ -14,10 +14,9 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 
 import lombok.RequiredArgsConstructor;
-import project.pickme.common.exception.BusinessException;
-import project.pickme.item.dto.ImageDto;
+import project.pickme.image.dto.ImageDto;
 import project.pickme.item.dto.ItemDto;
-import project.pickme.item.repository.ImageMapper;
+import project.pickme.image.repository.ImageMapper;
 
 @Service
 @RequiredArgsConstructor
