+++ pickme/src/main/resources/mapper/ImageMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="project.pickme.item.repository.ImageMapper">
+<mapper namespace="project.pickme.image.repository.ImageMapper">
 
     <insert id="insertImage">
         insert into Image(item_id, name, url, seq)
