+++ pickme/src/main/resources/mapper/ItemMapper.xml
@@ -1,32 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="project.pickme.item.repository.ItemMapper">
-
-    <resultMap id="oneItemResultMap" type="project.pickme.item.domain.Item">
-        <id property="id" column="item_id"/>
-        <result property="name" column="name"/>
-        <result property="target" column="target"/>
-        <result property="price" column="price"/>
-        <result property="startTime" column="start_time"/>
-        <result property="endTime" column="end_time"/>
-        <result property="status" column="status"/>
-        <association property="customs" column="customs_id" javaType="project.pickme.user.domain.Customs"/>
-    </resultMap>
-
-    <select id="findItemById" parameterType="long" resultMap="oneItemResultMap">
-        select i.item_id,
-               i.name,
-               i.target,
-               i.price,
-               i.start_time,
-               i.end_time,
-               i.status,
-               c.id as customs_id
-        from Item i
-                 left join Customs c on i.customs_id = c.id
-        where i.item_id = ${itemId}
-    </select>
-
     <insert id="insertItem" parameterType="ItemDto" useGeneratedKeys="true" keyProperty="itemId">
         insert into Item (name, code, target, price, start_time, end_time, status, customs_id)
         values (#{name}, #{code}, #{type}, #{price}, #{startTime}, #{endTime}, #{status}, #{customsId})
